[FEATURE] store double precision arrays

+++ src/openms/include/OpenMS/METADATA/DataArrays.h
@@ -44,7 +44,7 @@ namespace OpenMS
     /// Float data array class
     class FloatDataArray :
       public MetaInfoDescription,
-      public std::vector<float>
+      public std::vector<double>
     {
     };
